// src/app/models/solicitud-matricula.ts
export class SolicitudMatricula {

    id: number;
    estudiante: string;
    estudiante_id: number;
    celular: string;
    cedula: string;
    correo: string;
    carrera: string;
    carrera_id: number;
    ciclo: string;
    ciclo_id: number;
    periodo: string;
    periodo_id: number;
    tipo: string;
    estado: boolean;
    create_at: string;
    archivo: string;
    ruta: string;
    observaciones: string;
    fecha_modificacion: string;

}